test: Fixed hardcoded binary path in test
